caliperreader: handle root nodes in _create_parent
When creating parent nodes, we need to handle the case that the parent might be
a root node. Previously, the recursive _create_parent calls were being made on
root nodes, and we incorrectly tried to index into the grandparent callpath
tuple, even though it was empty. This ends the recursion if we encounter an
empty callpath tuple.