Read All PFW Files
The PerfFlowAspect Reader used to require that .pfw files
(PerfFlowAspect Traces) conform to JSON standard. However,
PerfFlowAspect does not generate files that are consistent with JSON.
This commit modifies a PerfFlowAspect trace file that does not conform
to JSON standard - by adding the closing brace and removing the trailing
comma.